requirements.txt
opencv-python
mediapipe
numpy
pandas
scikit-learn
pyttsx3
pillow